Summary: Fixes #22 - pipe support
**problem**
It would be nice to have the ability to pass options in via a pipe.

**analysis**
This is easy on all platforms, but the problem is that then you can't
have interactive input afterwards, which is kind of necessary.

On 'nix environments, you can re-attach to /dev/tty.

On windows environments, you might be able to connect to a COM.
https://stackoverflow.com/questions/19172563/how-do-i-accept-piped-input-and-then-user-prompted-input-in-a-python-script

I don't have any windows test environments, so I'm making it unix only,
since that's all I can verify.

**testing**
added one test for piping.
removed a test for showing the usage message when entering no args on
the cli, which now waits for piped args, and then yells about no items.

**documentation**
added documentation describing the piping feature in the CLI section.