Summary: Fixes #23 - tab completion
**Problem:**
No tab completion

**Analysis:**
Tab completion can be provided on *nix environments via the
readline library.  Use it, and add proper callback functions to handle
completion requests.

When performing completions, return the set of options which start with
the given text.

If case-insensitivity is enabled, case-insensitvely match the possible
options when completing.

If 'fuzzy' matching is enabled, match words according to the fuzzy rules.

Do not try to deal with 'search' matching - it's deprecated.

**Testing:**
Wrote tests for tab-completion, tab-completion with
insensitivity, and tab-completion with 'fuzzy' matching.

**Documentation:**
added some short docs for tab-completion