Summary: fixes #43 - empty items allowed
**Problem**
Empty items were allowed, which could never be matched.

**Analysis**
The item list wasn't being checked for empty items, just iterability and
finiteness, and non-emptiness (as a list).

Adding a check after deduplication fixes the immediate problem, but
leaves the similar issue of all-whitespace items and trailing-whitespace
items.

Using python's 'rstrip' prior to deduplication and empty-item checking
takes care of both of those related issues.

As with deduplication, the rstripping and the empty-item removal is done
automatically.  The assumption is that the user does not actually want
to display these things, as it is nonsensical, so it is *safe* to remove
them.  Further, it is asserted that if the user cares about catching
these mistakes themselves, they can do so easily prior to calling the
menu API.  If this becomes a common use case, it will be a new feature
to add a --pedantic or --no-assumptions flag, or something like that.

Also, something missed prior was default selection when deduplication is
occurring.  Prior to this commit, if deduplication caused the default
selection to be out-of-bounds, an error would be thrown.  Now the
default selection is processed prior to deduplication, so the expected
value is preserved.

**Testing**
Added 7 new tests to cover rstripping, deduplication after rstripping,
empty removal after rstripping, early default-selection, and selection
of an empty default.

**Documentation**
Added documentation for the rstripping, empty item removal, and early
default selection logic.