Summary: Fixes #44 - partial matches
**problem**
If an option is a subset of another option (like the following):
 * foo
 * foo bar

There is no way to select the partial option.

**analysis**
Add checks for exact matches with the options.  if the user response
exactly matches an option, use that option.

Add checks to verify this even in the case of fuzzy matching, but
restrict the definition of "exact" match to be exactly matching whole
words, not partial words, as the user may further elaborate to include
partials or words from the longer options.

**testing**
Added two new tests

**documentation**
as I believe this is the expected behavior, no new documentation is
needed.