Summary: Fixes #67 - incorrect tab completion matches
**Problem**
If there are multiple matches at the beginning of options, and also
options with the text in the middle or end, the tab completion match
list shows all of them, rather than just the matches which start with
the text.

**Analysis**
Allow the problematic behavior in fuzzy matching mode (where it makes
sense), but not in the standard mode.  Make sure in standard mode the
only matches displayed are the matches that start with the text.

**Testing**
Added a new test case.

**Documentation**
No new docs needed - bugfix.