added test for piping with tab, as well as code to support that
summary: fixes #70 - backspace break-on-fix

**problem**
The previous "fix" for the piping and tabbing issue sidestepped
input/raw_input on python3/2 by printing the prompt to the correct
output stream directly.

Unfortunately, this meant that the input functions didn't know not to
clear the line when the user backspaced past the first input
character... so the line got cleared (and the prompt disappeared), and
this is not the desired behavior.

**analysis**
The correct course of action is to set sys.stdout to be the desired
output stream, since the docs indicate that raw_input/input use
sys.stdout to print.

This causes other issues on python3, though, because python3's
implementation of input does not honor the reassignment of either stdin
or stdout, so tab completion is not functional if either of those
redirections are made.  issue #75 and issue #76, respectively, have
been filed in response.

**testing**
Added several new tests for tab completion and redirection, as well as
retrofitted old tests to deal with the python3 behavior now that the
streams are being redirected.

**documentation**
None so far, but the known errors should be documented.