Summary:  Fixes #74 - document the test cases
**Problem:**
Not much internal documentation of test cases

**Analysis:**
Many of the tests have reasonable documentation already, but many do
not.  Add enough documentation to each test as necessarey to make sure
it is readable.

**Testing:**
None, this is the documentation for the tests... there's no testing of
this.

**Documentation:**
No public documentation, because there is no public test documentation,
and I don't think there needs to be, but this commit itself is interal
documentation.