Summary:  Fixes #61 - remove deprecated features
**Problem:**
Deprecated features (search-mode and pre-prompt as the first arg to menu) were cluttering the code and tests.

**Analysis:**
Remove the deprecated features, the tests for them, and the
documentation for them.

**Testing:**
Removed the tests for search-mode.
Updated tests which used pre-prompt as the first arg to send the
pre-prompt as the second arg.

**Documentation:**
Removed the documentation for the deprecated features.