Summary:  Fixes #59 - runnable doc example
**Problem:**
The documented examples haven't been run in a while...

**Analysis:**
They really should have automatic tests somehow, but I'm not sure how to
do that at this point.  Re-run all of the documented examples by hand.

Found one typo in a prompt - fixed it.

**Testing:**
Unfortunately, manual

**Documentation:**
Updated the README to fix the typo.