Summary:  Fixes #82 - fuzzy matching broken
**Problem:**
Several failures in the fuzzy matching algorithm.

**Analysis:**
The fuzzy-matching algorithm was too greedy, and did not account for
things such as an early token matched several option tokens, but a later token ONLY
matched one option token, which was now gone.

Rewrite the algorithm to:
 - match partial input tokens in order of how many option tokens match
   (fewest first)
 - match option tokens in order of how many partial input tokens match
   (fewest first)

The first means, for instance, that if a partial input token (PIT) only
matches one option token (OT), that match is made first so that the matching
option token is not stolen by a more generic partial (PIT 'f' can't steal
OT 'food' from PIT 'foo').

The second means, for instance, that if a PIT matchs two OT's, but one
of those OT's is only matched by this PIT, that OT is taken, and the
other is left for some other PIT to match.

This is as optimal as I can get without trying every combination, I
think, so it is where I'm going to stop for now.

**Testing:**
Added a test for these kinds of combinations, which exposed several
problems.

**Documentation:**
Internal only.  This was a bug - the documented behavior is still the
behavior.